pyproject.toml
[tool.poetry]
name = "gyanasuchi"
version = "0.1.0"
description = "One who indexes knowledge"
authors = ["Karun Japhet <karunj@sahaj.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
youtube-transcript-api = "^0.6.1"
pytube = "^15.0.0"
modal = "^0.56.4326"
peewee = "^3.17.0"
pre-commit = "^3.6.0"
load-dotenv = "^0.1.0"
tqdm = "^4.66.1"
langchain = "^0.1.0"
qdrant-client = "^1.7.0"
uvicorn = "^0.25.0"
sentence-transformers = "^2.2.2"
openai = "^1.7.2"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.0"
langchain-community = "^0.0.12"
humanize = "^4.9.0"
pytest = "^7.4.4"
ragas = "^0.0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
